edit user details saves to repo #feature - still need to validate changes and retrieve current user from db - changed UserRepository#findById param to int

+++ identityprovider/src/main/java/nz/ac/canterbury/seng302/identityprovider/database/UserRepository.java
@@ -7,5 +7,5 @@ public interface UserRepository extends CrudRepository<UserModel, Long> {
 
   UserModel findByUsername(String username);
 
-  UserModel findById(long id);
+  UserModel findById(int id);
 }
